+++ src/win32/GDDataTranfer.h
