+++ src/win32/GetWord.c
