+++ src/win32/GetWord.h
