+++ src/win32/GetWordByIAccEx.c
