+++ src/win32/GetWordByIAccEx.h
