+++ src/win32/HookImportFunction.c
