+++ src/win32/TextOutHook.c
