+++ src/win32/TextOutHook.h
