+++ src/win32/TextOutSpy.c
