+++ src/win32/TextOutSpy.h
