+++ src/win32/ThTypes.c
