+++ src/win32/ThTypes.h
