+++ src/win32/guids.c
