+++ src/win32/readme.txt
