+++ src/win32/x64hooks.c
